Move CertUtils to `Conjur` module
